+++ nzilbb.labbcat.js
@@ -1419,9 +1419,9 @@
 		  var response = JSON.parse(responseText);
 		  result = response.model.result || response.model;
 		  errors = response.errors;
-		  if (errors.length == 0) errors = null
+		  if (errors && errors.length == 0) errors = null
 		  messages = response.messages;
-		  if (messages.length == 0) messages = null
+		  if (messages && messages.length == 0) messages = null
 		} catch(exception) {
 		  result = null
                   errors = ["" +exception+ ": " + labbcat.responseText];
@@ -2338,9 +2338,9 @@
 		  var response = JSON.parse(responseText);
 		  result = response.model.result || response.model;
 		  errors = response.errors;
-		  if (errors.length == 0) errors = null
+		  if (errors && errors.length == 0) errors = null
 		  messages = response.messages;
-		  if (messages.length == 0) messages = null
+		  if (messages && messages.length == 0) messages = null
 		} catch(exception) {
 		  result = null
                   errors = ["" +exception+ ": " + labbcat.responseText];
@@ -2594,9 +2594,9 @@
 		  var response = JSON.parse(responseText);
 		  result = response.model.result || response.model;
 		  errors = response.errors;
-		  if (errors.length == 0) errors = null
+		  if (errors && errors.length == 0) errors = null
 		  messages = response.messages;
-		  if (messages.length == 0) messages = null
+		  if (messages && messages.length == 0) messages = null
 		} catch(exception) {
 		  result = null
                   errors = ["" +exception+ ": " + labbcat.responseText];
@@ -2710,9 +2710,9 @@
 		  var response = JSON.parse(responseText);
 		  result = response.model.result || response.model;
 		  errors = response.errors;
-		  if (errors.length == 0) errors = null
+		  if (errors && errors.length == 0) errors = null
 		  messages = response.messages;
-		  if (messages.length == 0) messages = null
+		  if (messages && messages.length == 0) messages = null
 		} catch(exception) {
 		  result = null
                   errors = ["" +exception+ ": " + labbcat.responseText];
@@ -2788,6 +2788,252 @@
         "deleteParticipant", null, onResult, null, "POST",
         this.storeEditUrl, "application/x-www-form-urlencoded")
         .send(this.parametersToQueryString({id : id}));
+    }
+
+    /**
+     * Upload a transcript file and associated media files, as the first stage in adding or
+     * modifying a transcript to LaBB-CAT. The second stage is {@link #transcriptUploadParameters}
+     * @param {file|string} transcript The transcript to upload. In a browser, this
+     * must be a file object, and in Node, it must be the full path to the file. 
+     * @param {string|object} media The media to upload, if any. This can be a single
+     * file, or a map of  track suffixes to media files to upload for that track. In a
+     * browser, attribute values must be arrays of file objects, and in Node, they must be
+     * arrays of strings representing the full path to the file.
+     * @param merge Whether the upload corresponds to updates to an existing transcript
+     * (true) or a new transcript (false).
+     * @param {resultCallback} onResult Invoked when the request has returned a
+     * result, which is and object that hass the following attributes::
+     * <dl>
+     *  <dt> id </dt> <dd> The unique identifier to use for this upload when subsequently
+     *          calling {@link #transcriptUploadParameters}. </dd>
+     *  <dt> parameters </dt> <dd> An array of objects representing the parameters that
+     *          require values to be passed into {@link #transcriptUploadParameters}. 
+     *          The <q>parameters</q> returned may include both information
+     *          required by the format deserializer (e.g. mappings from tiers to LaBB-CAT
+     *          layers) and also general information required by LaBB-CAT (e.g. the
+     *          corpus, episode, and type of the transcript). 
+     *                 </dd>
+     * <p> Each parameter may contain the following attributes:
+     *  <dl>
+     *   <dt> name </dt>
+     *       <dd> The name that should be used when specifying the value for the parameter
+     *        when calling {@link #transcriptUploadParameters}. </dd> 
+     *   <dt> label </dt>
+     *       <dd> A label for the parameter intended for display to the user.</dd> 
+     *   <dt> hint </dt>
+     *       <dd> A description of the purpose of the parameter, for display to the user.</dd> 
+     *   <dt> type </dt>
+     *       <dd> The type of the parameter, e.g. <q>String</q>, <q>Double</q>, <q>Integer</q>,  
+     *         <q>Boolean</q>.</dd> 
+     *   <dt> required </dt>
+     *       <dd> <tt>true</tt> if the value must be specified, <tt>false</tt> if it is optional.</dd> 
+     *   <dt> value </dt>
+     *       <dd> A default value for the parameter.</dd> 
+     *   <dt> possibleValues </dt>
+     *       <dd> A list of possible values, if the possibilities are limited to a finite set.</dd> 
+     *  </dl>
+     * <p> The required parameters may include both information
+     * required by the format deserializer (e.g. mappings from tiers to LaBB-CAT layers) 
+     * and also general information required by LaBB-CAT, such as:
+     *  <dl>
+     *   <dt> labbcat_corpus </dt>
+     *       <dd> The corpus the new transcript(s) belong(s) to. </dd> 
+     *   <dt> labbcat_episode </dt>
+     *       <dd> The episode the new transcript(s) belong(s) to. </dd> 
+     *   <dt> labbcat_transcript_type </dt>
+     *       <dd> The transcript type for the new transcript(s). </dd> 
+     *   <dt> labbcat_generate </dt>
+     *       <dd> Whether to re-regenerate layers of automated annotations or not. </dd> 
+     *  </dl> 
+     * @param onProgress Invoked on XMLHttpRequest progress.
+     */
+    transcriptUpload(transcript, media, merge, onResult, onProgress) {
+      if (typeof media === "boolean") {
+        // (transcript, merge, onResult, onProgress)
+        onProgress = onResult;
+        onResult = merge;
+        merge = media;
+        media = null;
+      }
+      if (exports.verbose) {
+        console.log("transcriptUpload(" + transcript + ", " + media + ", " + merge + ")");
+      }
+      // create form
+      var fd = new FormData();
+      fd.append("merge", ""+merge);
+      
+      if (!runningOnNode) {	
+        
+	fd.append("transcript", transcript);
+	if (media) {
+          if (typeof media != "object") media = { "" : [media] }; // convert to track->array map
+          for (var trackSuffix in Object.keys(media)) {
+            var files = media[trackSuffix];
+	    if (files === Array) { // multiple files
+	      for (var f in files) {
+	        fd.append("media"+trackSuffix, files[f]);
+	      } // next file
+            } else { // a single file
+	      fd.append("media"+trackSuffix, files);
+	    }
+          } // next track suffix
+	} // media to upload
+        
+	// create HTTP request
+	var xhr = new XMLHttpRequest();
+	xhr.call = "transcriptUpload";
+	xhr.id = transcript.name;
+	xhr.onResult = onResult;
+	xhr.addEventListener("load", callComplete, false);
+	xhr.addEventListener("error", callFailed, false);
+	xhr.addEventListener("abort", callCancelled, false);
+	xhr.upload.addEventListener("progress", onProgress, false);
+	xhr.upload.id = transcript.name; // for knowing what status to update during events
+	
+	xhr.open("POST", this.baseUrl + "api/edit/transcript/upload");
+	if (this.username) {
+	  xhr.setRequestHeader("Authorization", "Basic " + btoa(this.username + ":" + this.password))
+	}
+	xhr.setRequestHeader("Accept", "application/json");
+	xhr.send(fd);
+      } else { // runningOnNode
+	
+	// on node.js, files are actually paths
+	var transcriptName = transcript.replace(/.*\//g, "");
+        if (exports.verbose) console.log("transcriptName: " + transcriptName);
+
+	fd.append(
+          "transcript", 
+	  fs.createReadStream(transcript).on('error', function(){
+	    onResult(
+              null, ["Invalid transcript: " + transcriptName], [], "transcriptUpload", transcriptName);
+	  }), transcriptName);
+        
+        if (media) {
+          if (typeof media == "string") media = { "" : [media] }; // convert to track->array map
+          for (var trackSuffix in media) {
+            var files = media[trackSuffix];
+	    if (files.constructor === Array) { // multiple files
+	      for (var f in files) {
+	        var mediaName = files[f].replace(/.*\//g, "");
+	        try {
+		  fd.append(
+                    "media"+trackSuffix, 
+		    fs.createReadStream(files[f]).on('error', function(x){
+		      onResult(
+                        null, ["Invalid media: " + mediaName], [], "transcriptUpload", transcriptName);
+		    }), mediaName);
+	        } catch(error) {
+		  onResult(
+                    null, ["Invalid media: " + mediaName, error.code], [], "transcriptUpload", 
+                    transcriptName);
+		  return;
+	        }
+	      } // next file
+            } else { // a single file
+	      var mediaName = files.replace(/.*\//g, "");
+	      try {
+	        fd.append(
+                  "media"+trackSuffix, 
+		  fs.createReadStream(files).on('error', function(){
+		    onResult(null, ["Invalid media: " + mediaName], [], "transcriptUpload", transcriptName);
+		  }), mediaName);
+	      } catch(error) {
+		onResult(
+                  null, ["Invalid media: " + mediaName, error.code], [], "transcriptUpload", 
+                  transcriptName);
+		return;
+	      }
+	    } // single file
+	  } // next track suffix
+	  
+	  var urlParts = parseUrl(this.baseUrl + "api/edit/transcript/upload");
+	  // for tomcat 8, we need to explicitly send the content-type and content-length headers...
+	  var labbcat = this;
+          var password = this._password;
+	  fd.getLength(function(something, contentLength) {
+	    var requestParameters = {
+	      port: urlParts.port,
+	      path: urlParts.pathname,
+	      host: urlParts.hostname,
+	      headers: {
+	        "Accept" : "application/json",
+	        "content-length" : contentLength,
+	        "Content-Type" : "multipart/form-data; boundary=" + fd.getBoundary()
+	      }
+	    };
+	    if (labbcat.username && password) {
+	      requestParameters.auth = labbcat.username+':'+password;
+	    }
+	    if (/^https.*/.test(labbcat.baseUrl)) {
+	      requestParameters.protocol = "https:";
+	    }
+            if (exports.verbose) {
+              console.log("submit: " + labbcat.baseUrl + "api/edit/transcript/upload");
+            }
+	    fd.submit(requestParameters, function(err, res) {
+	      var responseText = "";
+	      if (!err) {
+	        res.on('data',function(buffer) {
+		  responseText += buffer;
+	        });
+	        res.on('end',function(){
+	          var result = null;
+	          var errors = null;
+	          var messages = null;
+		  try {
+		    var response = JSON.parse(responseText);
+		    result = response.model.result || response.model;
+		    errors = response.errors;
+		    if (errors && errors.length == 0) errors = null
+		    messages = response.messages;
+		    if (messages && messages.length == 0) messages = null
+		  } catch(exception) {
+		    result = null
+                    errors = ["" +exception+ ": " + labbcat.responseText];
+                    messages = [];
+		  }
+		  onResult(result, errors, messages, "transcriptUpload", transcriptName);
+	        });
+	      } else {
+	        onResult(null, ["" +err+ ": " + labbcat.responseText], [], "transcriptUpload", transcriptName);
+	      }
+	      
+	      if (res) res.resume();
+	    });
+	  }); // got length
+        } // media is specified
+      } // runningOnNode
+    } // transcriptUpload
+    
+    /**
+     * The second part of a transcript upload process started by a call to
+     * {@link #transcriptUpload}, which specifies values for the parameters
+     * required to save the uploaded transcript to LaBB-CAT's database. 
+     * <p> If the response includes more parameters, then this method should be called again
+     * to supply their values.
+     * @param {string} id Upload ID returned by the prior call to {@link #transcriptUpload}.
+     * @param {object} parameters Object with an attribute and value for each parameter
+     * returned by the prior call to {@link #transcriptUpload}.
+     * @param {resultCallback} onResult Invoked when the request has returned a
+     * result, which is and object that hass the following attributes::
+     * <dl>
+     *  <dt> transcripts </dt> <dd> an object for which each key is a transcript name, and its 
+     *          value is the threadId of the server task processing the uploaded transcript, 
+     *          which can be passed to {@link LabbcatView#taskStatus} to monitor progress. </dd>
+     *  <dt> id </dt> <dd> The unique identifier to use for this upload when subsequently
+     *          calling {@link #transcriptUploadParameters} (if parameters are returned). </dd>
+     *  <dt> parameters </dt> <dd> An array of objects representing the parameters that
+     *          still require values. </dd>
+     * <p> If parameters are returned, they have the same structure as those returned by 
+     * {@link #transcriptUpload}
+     */
+    transcriptUploadParameters(id, parameters, onResult) {
+      this.createRequest(
+        "transcriptUploadParameters", null, onResult,
+        this.baseUrl+"api/edit/transcript/upload/"+encodeURIComponent(id)
+          + "?"+this.parametersToQueryString(parameters),
+        "PUT").send();
     }
 
     /**
@@ -2942,9 +3188,9 @@
 		  var response = JSON.parse(responseText);
 		  result = response.model.result || response.model;
 		  errors = response.errors;
-		  if (errors.length == 0) errors = null
+		  if (errors && errors.length == 0) errors = null
 		  messages = response.messages;
-		  if (messages.length == 0) messages = null
+		  if (messages && messages.length == 0) messages = null
 		} catch(exception) {
 		  result = null
                   errors = ["" +exception+ ": " + labbcat.responseText];
@@ -3051,9 +3297,9 @@
 		var response = JSON.parse(responseText);
 		result = response.model.result || response.model;
 		errors = response.errors;
-		if (errors.length == 0) errors = null
+		if (errors && errors.length == 0) errors = null
 		messages = response.messages;
-		if (messages.length == 0) messages = null
+		if (messages && messages.length == 0) messages = null
                 ;
 	      } catch(exception) {
 		result = null
@@ -3984,9 +4230,9 @@
 		  var response = JSON.parse(responseText);
 		  result = response.model.result || response.model;
 		  errors = response.errors;
-		  if (errors.length == 0) errors = null
+		  if (errors && errors.length == 0) errors = null
 		  messages = response.messages;
-		  if (messages.length == 0) messages = null
+		  if (messages && messages.length == 0) messages = null
 		} catch(exception) {
 		  result = null
                   errors = ["" +exception+ ": " + labbcat.responseText];
@@ -4206,9 +4452,9 @@
 		  var response = JSON.parse(responseText);
 		  result = response.model.result || response.model;
 		  errors = response.errors;
-		  if (errors.length == 0) errors = null
+		  if (errors && errors.length == 0) errors = null
 		  messages = response.messages;
-		  if (messages.length == 0) messages = null
+		  if (messages && messages.length == 0) messages = null
 		} catch(exception) {
 		  result = null
                   errors = ["" +exception+ ": " + labbcat.responseText];